+++ src/api/instance/authInstance.js
@@ -1,0 +1,24 @@
+import axios from 'axios';
+
+const authAPI = (url, options) => {
+  return axios.create({ baseURL: url, ...options });
+};
+
+const authInstance = authAPI(process.env.REACT_APP_BASE_URL);
+
+authInstance.interceptors.request.use(
+  (config) => {
+    const token = localStorage.getItem('token');
+    console.log(token);
+    config.headers['Content-Type'] = 'application/json';
+    config.headers['Authorization'] = `Bearer ${token}`;
+
+    return config;
+  },
+  (error) => {
+    console.log(error);
+    return Promise.reject(error);
+  },
+);
+
+export default authInstance;
